requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cssselect==1.1.0
defusedxml==0.7.1
Django==4.0.5
django-admin-interface==0.19.2
django-ckeditor==6.4.2
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==2.0.0
django-phonenumber-field==6.3.0
django-restframework==0.0.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.1
markdown2==2.4.3
MarkupSafe==2.1.1
oauthlib==3.2.0
phonenumbers==8.12.53
Pillow==9.2.0
pycparser==2.21
PyJWT==2.4.0
pyquery==1.4.3
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
readtime==2.0.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.11
whitenoise==6.2.0
gunicorn
